Move status deletion logic back to statuses.js
